requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
typer[all]
jsonschema
